requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-ipware==7.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.5.0
geoip2==5.0.1
gunicorn==23.0.0
hiredis==3.1.0
idna==3.10
kombu==5.5.2
maxminddb==2.6.3
multidict==6.1.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.51
propcache==0.3.0
psycopg2-binary==2.9.10
pyaes==1.6.1
pycparser==2.22
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==3.0.0
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3